all scripts in python
