+++ src/safety_clean.py
@@ -1,0 +1,40 @@
+import subprocess
+from config import findora_env
+from shared import (
+    stop_and_remove_container,
+    chown_dir,
+    get_live_version,
+    create_local_node,
+    get_snapshot,
+)
+
+
+def run_safety_clean(network, region):
+    USERNAME = findora_env.active_user_name
+
+    ENV = "prod"
+
+    server_url = f"https://{ENV}-{network}.{ENV}.findora.org"
+
+    LIVE_VERSION = get_live_version(server_url)
+
+    FINDORAD_IMG = f"findoranetwork/findorad:{LIVE_VERSION}"
+    CHECKPOINT_URL = f"https://{ENV}-{network}-us-west-2-ec2-instance.s3.us-west-2.amazonaws.com/{network}/checkpoint"
+    ROOT_DIR = f"/data/findora/{network}"
+    keypath = f"{ROOT_DIR}/{network}_node.key"
+    CONTAINER_NAME = "findorad"
+
+    chown_dir(ROOT_DIR, USERNAME, USERNAME)
+
+    stop_and_remove_container(CONTAINER_NAME)
+
+    # get snapshot
+    get_snapshot(ENV, network, ROOT_DIR, region)
+
+    # get checkpoint on testnet
+    if network == "testnet":
+        subprocess.run(["sudo", "rm", "-rf", f"{ROOT_DIR}/checkpoint.toml"], check=True)
+        subprocess.run(["wget", "-O", f"{ROOT_DIR}/checkpoint.toml", f"{CHECKPOINT_URL}"], check=True)
+
+    # Start findorad
+    create_local_node(ROOT_DIR, FINDORAD_IMG)
